+++ packages/next/src/index.ts
@@ -1,7 +1,0 @@
-export * from './client';
-export * from './config';
-export * from './getProps';
-export * from './HeadlessProvider';
-export * from './utils';
-export * from './withFaust';
-export * from './log';